pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hafix-agent"
version = "0.1.0"
description = "HAFixAgent: History-Aware Automated Program Repair Agent"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "HAFixAgent"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.10"

# Core dependencies - using compatible versions
dependencies = [
    "mini-swe-agent>=1.12.0",  # Main dependency for SWE agent functionality
    "datasets>=4.0.0",
    "litellm>=1.75.9",
    "pyyaml>=6.0.2",
    "openai>=1.99.5",
    "pillow>=10.0.0",  # More compatible version
    "pytz>=2024.1",    # More compatible version
    "jinja2>=3.1.0",
    "numpy>=1.21.0",   # More compatible version
    "packaging>=21.0",
    "scipy>=1.7.0",    # More compatible version
    "certifi>=2023.0.0",   # More compatible version
    "fsspec>=2023.1.0,<=2025.3.0",  # Compatible with datasets
    "six>=1.16.0",
    "mpmath>=1.2.0",
    "pandas>=1.3.0",   # More compatible version
    "rich>=12.0.0",
    "python-dotenv>=0.19.0",
    "typer[all]>=0.12.0",
    "textual>=0.1.0",
    "requests>=2.25.0",
    "tenacity>=8.0.0",
    "platformdirs>=2.0.0",
    "ghapi",
    "beautifulsoup4",
    "javalang",
    "matplotlib>=3.4.0",
    "matplotlib-venn>=1.1.0",
    "seaborn>=0.11.0",
    "tiktoken>=0.5.0",  # For token counting in analysis scripts
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/ThomasShiyu/HAFixAgent"
Repository = "https://github.com/ThomasShiyu/HAFixAgent"
Issues = "https://github.com/ThomasShiyu/HAFixAgent/issues"

[project.scripts]
hafixagent = "evaluation.run_defects4j_evaluation:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["hafix_agent*", "dataset*", "evaluation*", "analysis*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json", "*.md"]

# Development tools configuration
[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
